test(outlook-semantic-mcp): fix sync-delegated-access tests after query extraction refactor
