+++ services/outlook-semantic-mcp/src/features/delegated-access/utils/is-delegated-access-not-available-error.ts
